requirements.txt
datasets==2.18.0
matplotlib==3.10.5
pandas==2.3.1
PyYAML==6.0.2
PyYAML==6.0.2
tensorboard==2.19.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
